requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
cryptography==45.0.2
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
numpy==1.26.4
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pluggy==1.6.0
prisma==0.15.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tabulate==0.9.0
tomlkit==0.13.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
